wip: initial sketch for rogga api?
This needs a lot more work and I might need
to rewrite other things, like the package-arg crate,
but I am starting to like the direction this is all
going? Just needs a bit more tweaking.

The biggest challenge is realizing that I want an API
where resolvers can be injected for specific resolution
logic, as opposed to the biggest number wins (with caveats)
approach that NPM (currently) takes.