Compress cache contents using zstd
Use zstd to compress cache contents on write and decompress on read.
I chose zstd because of its pluggable compression dictionary, which
can be trained on a sample corpus to improve its compression and
decompression ratio and speed. I haven't wired this up yet, but I'm
interested in using popular npm packages as the corpus.

Further improvements can be made to the copy/copy_hash functions if
we record expected decompressed size ahead of the stream: we could
copy using mmap!

BREAKING CHANGE: This change makes the Reader's check() function
asynchronous.