Add logo assets.
I originally made the logo in procreate, but exported in PSD and PNG formats.
All layers should be preserved in the procreate and PSD files, so please don't
judge me for my "Layer 1...N" 😅 The PSD and procreate exports have a solid
background layer, but I exported the PNG on a transparent background.

I intend to follow this up with an SVG version, but if someone beats me to the
punch that's totally okay!